# api-spec/openapi.yaml
openapi: "3.1.0"
info:
  title: imajin events
  version: "1.0.0"
  description: >
    Event creation, ticketing, ticket tiers, queue management, hold system, .fair manifests, checkout, and payment webhooks.
    Served by the app itself at /api/spec. Authenticate with a scoped app token (`Authorization: Bearer`, audience `events`,
    scopes `events:read` for GET/HEAD/OPTIONS and `events:write` otherwise) or, during migration, the legacy session cookie.
servers:
  - url: https://events.imajin.ai
    description: production
  - url: https://dev-events.imajin.ai
    description: dev
  - url: http://localhost:3000
    description: local
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: imajin_session
    bearerAuth:
      type: http
      scheme: bearer
      description: "Scoped app token minted for audience `events` (events:read for safe methods, events:write otherwise)"
    webhookSecret:
      type: http
      scheme: bearer
      description: "Bearer WEBHOOK_SECRET — internal payment webhook from pay service"
  responses:
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
paths:
...
  /api/events/{id}/tickets/{ticketId}/refund:
    post:
      tags: [events]
      operationId: refundTicket
      summary: Mark a ticket as refunded — event owner only
      security:
        - cookieAuth: []
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: ticketId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Ticket refunded
          content:
            application/json:
              schema:
                type: object
                properties:
                  ticket:
                    type: object
                    properties:
                      id: { type: string }
                      status: { type: string }
        "400": { description: Ticket not in refundable state, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { description: Only event owner can issue refunds, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
        "404": { description: Ticket or event not found, content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } } }
...